Revert "Gutenberg blocks for IMPress lead signup and login widgets"
